// src/region/region-name.enum.ts
export enum RegionName {
    TASHKENT = 'Toshkent',
    SAMARKAND = 'Samarqand',
    ANDIJAN = 'Andijon',
    FERGANA = 'Farg‘ona',
    NAMANGAN = 'Namangan',
    BUKHARA = 'Buxoro',
    KHOREZM = 'Xorazm',
    KASHKADARYA = 'Qashqadaryo',
    SURKHANDARYA = 'Surxondaryo',
    JIZZAKH = 'Jizzax',
    NAVOIY = 'Navoiy',
    SYRDARYA = 'Sirdaryo',
  }
  